chore(compose): the vault option gets its own file — server.go back under the cap
